fix: consolidate legacy and default plugin and fix resizing
